move groups methods into new file
